+++ include/mtd/mtd-abi.h
@@ -8,7 +8,7 @@
 #define __MTD_ABI_H__
 
 struct erase_info_user {
-	uint32_t start;
+	uint64_t start;
 	uint32_t length;
 };
 
@@ -57,7 +57,7 @@ struct mtd_oob_buf {
 struct mtd_info_user {
 	uint8_t type;
 	uint32_t flags;
-	uint32_t size;	 // Total size of the MTD
+	uint64_t size;	 // Total size of the MTD
 	uint32_t erasesize;
 	uint32_t writesize;
 	uint32_t oobsize;   // Amount of OOB data per block (e.g. 16)
